Distribution of Pearson Correlation
We simple describe Pearson correlation as the measure of linear relationship between two continuous variables.   The coefficient also describes both the strength and the direction or sense of the relationship.
The graph clearly show that the distribution of the correlation is student t-distribution which tends to be normal as the sample size increases.